requirements.txt
llama-cpp-python
fastapi
uvicorn
langchain
chromadb
sentence-transformers
